requirements.txt
Django>=3.2.14,<4
pytz==2022.7.1
httplib2==0.22.0
feedparser==6.0.10
Markdown==3.4.3 # pyup: < 3
psycopg2==2.9.5
versiontools==1.9.1
statsd==4.0.1
pep8==1.7.1
pyflakes==3.0.1
mccabe==0.7.0
configparser==5.3.0
flake8==6.0.0
pycodestyle==2.10.0
typing_extensions==4.5.0
importlib-metadata<6.2  # for flake8
zipp==3.15.0  # for flake8
tornado==6.2
beautifulsoup4==4.12.0
cssselect==1.2.0
lxml==4.9.2
fuzzywuzzy==0.18.0
sure==2.0.1
coverage==7.2.2
pyasn1==0.4.8
cryptography==40.0.0  # pyOpenSSL
pyOpenSSL==23.1.0
ndg-httpsclient==0.5.1
urllib3==1.26.15  # requests, sentry-sdk
chardet==5.1.0  # requests
idna==3.4  # requests
certifi==2022.12.7  # requests, sentry-sdk
requests==2.28.2
python-cas==1.6.0
git+https://github.com/django-cas-ng/django-cas-ng.git

astroid==2.15.0
pylint==2.17.1

six==1.16.0
ipaddress==1.0.23  # faker
faker==18.3.1  # factory_boy
factory_boy==3.2.1

ldap3==2.9.1
sqlparse==0.4.3

# boto3 deps
s3transfer==0.6.0
jmespath==1.0.1

botocore>=1.12.198,<1.30.0
boto3>=1.9.198,<1.27.0

python-dateutil==2.8.2

django-indexer==0.3.0
django-templatetag-sugar==1.0
django-paging==0.2.5
django-annoying==0.10.6
django-appconf==1.0.5
rcssmin==1.1.1  # django-compressor
rjsmin==1.2.1  # django-compressor
django-compressor>=3.1,<5.0
django-statsd-mozilla==0.4.0
raven==6.10.0  # Remove from ccnmtlsettings, then here.
sentry-sdk==1.17.0
django-bootstrap4==22.3
django-debug-toolbar==3.8.1
django-waffle==3.0.0

django-smoketest==1.2.0
# For django-extensions
typing==3.7.4.1; python_version < '3.5'
django-extensions==3.2.1
django-stagingcontext==0.1.0
django-ga-context==0.1.0
django-impersonate==1.9.1
django-markwhat==1.6.2
gunicorn==20.1.0
django-infranil==1.1.0
django-flatblocks==1.0.0

djangorestframework==3.14.0

django-storages==1.13.2
ccnmtlsettings==1.9.5

django-reversion==5.0.4
text_unidecode==1.3

django-taggit==3.1.0

django-cacheds3storage==0.3.0

nameparser==1.1.2
oauth2==1.9.0.post1
oauthlib==3.2.2
pylti==0.7.0
django-lti-provider==1.0.0

pyparsing==3.0.9
edtf==4.0.1 # pyup: <4.0.0

rich==13.3.2 # bandit
pbr==5.11.1 # bandit
pyyaml==6.0 # bandit
stevedore==5.0.0  # bandit
bandit==1.7.5

django-smtp-ssl==1.0
django-contact-us==1.1.0

entrypoints==0.4

django-courseaffils==2.3.0

mock==5.0.1
asgiref==3.6.0
